Separate LyvEffectiveModelContext
Separate LyvEffectiveModelContext and create dedicated
LyvEffectiveModelContextFactory class.
