+++ src/test/java/io/lighty/yang/validator/check/update/from/RFC6020Test.java
@@ -7,12 +7,12 @@
  */
 package io.lighty.yang.validator.check.update.from;
 
-import static io.lighty.yang.validator.Main.getLyvContext;
 import static io.lighty.yang.validator.Main.runLYV;
 import static org.testng.Assert.assertEquals;
 
 import com.google.common.collect.ImmutableList;
 import io.lighty.yang.validator.Cleanable;
+import io.lighty.yang.validator.LyvEffectiveModelContextFactory;
 import io.lighty.yang.validator.Main;
 import io.lighty.yang.validator.config.Configuration;
 import io.lighty.yang.validator.config.ConfigurationBuilder;
@@ -107,7 +107,8 @@ public class RFC6020Test implements Cleanable {
                 .setYangModules(Collections.singletonList(newFile))
                 .build();
 
-        final var lyvContext = getLyvContext(ImmutableList.of(Paths.get(newFile).toString()), config);
+        final var lyvContext = LyvEffectiveModelContextFactory.create(
+                ImmutableList.of(Paths.get(newFile).toString()), config);
         final var modules = lyvContext.testedModules();
         assertEquals(modules.size(), 1);
         runLYV(modules.iterator().next(), config, null, lyvContext.context());
