+++ src/test/java/io/lighty/yang/validator/formats/JsTreeTest.java
@@ -7,12 +7,12 @@
  */
 package io.lighty.yang.validator.formats;
 
-import static io.lighty.yang.validator.Main.getLyvContext;
 import static io.lighty.yang.validator.Main.runLYV;
 import static org.testng.Assert.assertEquals;
 
 import com.google.common.collect.ImmutableList;
 import io.lighty.yang.validator.FormatTest;
+import io.lighty.yang.validator.LyvEffectiveModelContextFactory;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -36,7 +36,7 @@ public class JsTreeTest extends FormatTest {
         setFormat();
         final String module = Paths.get(yangPath).resolve("undeclared.yang").toString();
         final var configuration = builder.build();
-        final var lyvContext = getLyvContext(ImmutableList.of(module), configuration);
+        final var lyvContext = LyvEffectiveModelContextFactory.create(ImmutableList.of(module), configuration);
         final var modules = lyvContext.testedModules();
         assertEquals(modules.size(), 1);
         runLYV(modules.iterator().next(), configuration, formatter, lyvContext.context());
